refactor navigation site-data
